Moved tests to unit_tests/dagmc/test_model.py
